requirements.txt
Django>=4.2
djangorestframework
django-extensions>=3.1.3
drf-spectacular
Pillow